Prepare project for production
