+++ package.json
@@ -3,10 +3,6 @@
   "version": "1.0.0",
   "description": "",
   "main": "main.js",
-  "scripts": {
-    "start": "node ./server.js",
-    "dev": "nodemon ./server.js"
-  },
   "repository": {
     "type": "git",
     "url": "git@github.com:EventService/public-web.git"
@@ -17,6 +13,44 @@
     "express": "4.18.2"
   },
   "devDependencies": {
-    "nodemon": "2.0.22"
+    "autoprefixer": "10.4.14",
+    "concurrently": "8.2.0",
+    "cssnano": "6.0.1",
+    "nodemon": "2.0.22",
+    "npm-watch": "0.11.0",
+    "postcss": "8.4.24",
+    "postcss-cli": "10.1.0",
+    "rollup": "3.25.1",
+    "rollup-plugin-commonjs": "10.1.0",
+    "rollup-plugin-node-resolve": "5.2.0",
+    "rollup-plugin-terser": "7.0.2"
+  },
+  "watch": {
+    "minify-css": {
+      "patterns": [
+        "web"
+      ],
+      "extensions": [
+        "css"
+      ],
+      "delay": 250
+    },
+    "minify-js": {
+      "patterns": [
+        "web"
+      ],
+      "extensions": [
+        "js"
+      ],
+      "delay": 250
+    }
+  },
+  "scripts": {
+    "start-server": "nodemon ./server.js",
+    "minify-js": "rollup -c",
+    "minify-css": "postcss --no-map web/main.css > public/main.min.css",
+    "minify": "yarn minify-js && yarn minify-css",
+    "watch": "npm-watch",
+    "start": "concurrently \"yarn start-server\" \"yarn watch\""
   }
 }
