+++ postcss.config.mjs
@@ -1,0 +1,13 @@
+import autoprefixer from "autoprefixer";
+import cssnanoPlugin from "cssnano";
+
+export default {
+    plugins: [
+        autoprefixer({
+            preset: "default",
+        }),
+        cssnanoPlugin({
+            preset: "default",
+        }),
+    ],
+};
