+++ src/pages/news.html
@@ -18,7 +18,7 @@
 	<link rel="icon" type="image/png" sizes="32x32" href="/img/favicons/favicon-32x32.png">
 	<link rel="icon" type="image/png" sizes="96x96" href="/img/favicons/favicon-96x96.png">
 	<link rel="icon" type="image/png" sizes="16x16" href="/img/favicons/favicon-16x16.png">
-	<link href="/main.css" rel="stylesheet">
+	<link href="/main.min.css" rel="stylesheet">
 	<link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Inter:wght@400;500;600;700&display=swap"
 		rel="stylesheet">
 </head>
@@ -80,11 +80,11 @@
     </div>
   </footer>
 
-	<script src="/main.js" charset="utf-8"></script>
+	<script src="/main.min.js" charset="utf-8"></script>
 	<script>
 		document.addEventListener('DOMContentLoaded', function (event) {
 			//the event occurred         
-			getPostList()
+			main.getPostList()
 		})
 	</script>
 </body>
